interiordesign/requirements.txt
OpenAI
streamlit
python-dotenv
